+++ src/commands/requests/editBody.ts
@@ -1,6 +1,6 @@
 import { join } from 'path';
 import { Uri, window, workspace } from 'vscode';
-import { getCollection } from '../../utils';
+import { save } from '../../utils';
 import { PostmanItemModel } from '../../views/postmanItems/postmanItemModel';
 import { runCommand } from '../commands';
 
@@ -24,8 +24,7 @@ export async function editBody(item: PostmanItemModel): Promise<void> {
     if (e.document.uri === document.uri) {
       object.body = document.getText();
 
-      const collection = getCollection(object);
-      e.waitUntil(runCommand('saveCollection', collection));
+      e.waitUntil(save(object));
     }
   });