+++ src/commands/requests/editSettings.ts
@@ -1,16 +1,17 @@
-import { getCollection } from '../../utils';
+import { save } from '../../utils';
 import { PostmanItemModel } from '../../views/postmanItems/postmanItemModel';
 import { RequestSettingsView } from '../../views/requestSettingsView/requestSettingsView';
-import { runCommand } from '../commands';
 
 export function editSettings(item?: PostmanItemModel): void {
   if (item === undefined || !item.isRequest()) {
     return;
   }
 
-  new RequestSettingsView(item.itemObject.settings ?? {}, `${item.itemObject.name} Settings`)
+  const object = item.itemObject;
+
+  new RequestSettingsView(object.settings ?? {}, `${object.name} Settings`)
     .onChange((e) => {
-      item.itemObject.settings = e;
-      runCommand('saveCollection', getCollection(item.itemObject));
+      object.settings = e;
+      save(object);
     });
 }