+++ src/commands/requests/editUrl.ts
@@ -1,8 +1,7 @@
 import { Url } from 'postman-collection';
 import { window } from 'vscode';
-import { getCollection } from '../../utils';
+import { save } from '../../utils';
 import { PostmanItemModel } from '../../views/postmanItems/postmanItemModel';
-import { runCommand } from '../commands';
 
 export async function editUrl(item?: PostmanItemModel): Promise<void> {
   if (item === undefined || !item.isRequest()) {
@@ -16,5 +15,5 @@ export async function editUrl(item?: PostmanItemModel): Promise<void> {
   }
 
   request.url = new Url(url);
-  await runCommand('saveCollection', getCollection(request));
+  await save(request);
 }