+++ src/utils/index.ts
@@ -3,4 +3,5 @@ export * from './getCollections';
 export * from './getNonce';
 export * from './resolveChildren';
 export * from './runWithNewman';
+export * from './save';
 export * from './typeChecks';